adds_function_relpath, docs, tests
This adds function relpath which has similar functionalities as python's
os.path.relpath

Return a relative filepath to path either from the current directory or
from an optional start directory. This is a path computation: the
filesystem is not accessed to confirm the existence or nature of path or
start.